Add adt loading (#5)
* Change entity structure

* Add no support disclaimer and WIP in README

* Add appveyor

* Fix chuck namespace

* Add ubuntu build to appveyor

* Change Chucks to Chunks

* Fix chunk namespace

* Add MFBO and MTFX chunk

* Fix MDDF, MHDR, MODF, MWMO, MCIN chunks

* Change flag names

* Add ADT load logic

* Add wotlk adt class

* Fix MHDR and Rotator reader

* Add ADT writer

* Add optional chunk custom attribute

+++ ADTConvert2/Files/ADT/Flags/MHDRFlags.cs
@@ -6,7 +6,7 @@ namespace ADTConvert2.Files.ADT.Flags
     /// Flags for the <see cref="MHDRFlags"/>.
     /// </summary>
     [Flags]
-    public enum MHDRFlags : ushort
+    public enum MHDRFlags : uint
     {
         /// <summary>
         /// Map contains MFBO chuck
